introduce initial type structure
This commit adds the general type structure.
It does not add any implementation but only
an first suggestion on how the API should look
like.

The main idea is that the `aead::Algorithm` (as trait)
can be implemented by other crates such that different
implementations can be plugged into `sio`.