restructure crate types
This commit completely redefines the API of the `sio` crate.
The core type is `Algorithm` which now uses associated constants
for the key, nonce and tag length. Using ass. constants the compiler
can guarantee a lot of invariants, like:
 - The key passed to `Algorithm::new` has always the correct length.
   (Therefore we can return `Self` instead of a Result)
 - The nonce passed to `seal_in_place` / `open_in_place` has always
   the correct length.
 - It's actually not possible to implement `Algorithm` for an AEAD
   with a nonce size < 4 and use it with `sio`. Doing so causes the
   compiler to complain that `Nonce::<Algorithm>::SIZE` will overflow
   (usize becomes negative).

Further, it's not possible to pass a wrong key or nonce to an
enc/dec reader/writer.

However, this commit only contains the basic structure. A lot of things
are still missing. In particular:
  - docs and examples!!!
  - the enc/dec reader implementations
  - proper cargo.toml configuration (production/dev builds a.s.o)
  - tests and benchmarks.