make EncWriter implementation robust and add docs
This commit makes the EncWriter implementation robust
against:
  - an unexpected panic of the inner writer
  - misbehaving calling code that invokes e.g. write
    multiple times even though an error has occurred.
  - code that explicitly completes the encryption process
    but then tries to use the EncWriter again.

For panics the EncWriter has a boolean panic flag that is checked
by the Drop implementation to not try to close the EncWriter. After
a panic, the state (i.e. how many bytes have been written out) is
undefined.

When an error occurs the `EncWriter` sets the `errored` flag to true.
Before performing another write/flush/close it will check whether an
error has occurred in the past and fails if so.

Finally, a `EncWriter` must be closed (by calling `close`) to complete
the encryption process. If not done explicitly, it is closed during
dropping. Now, the `close` method consumes the `EncWriter` to prevent
calling code from using it later on.

Additionally this commit adds documentation for the public API of the
`EncWriter`.