refactor {Enc/Dec}Writer to use Box<[u8]> instead of Vec<u8>
This commit replaces the buffer in {Enc/Dec}Writer with a
Box<[u8]>. The reason is that non-sophisticated benchmarks
showed that truncating the Vec after a write to the inner
writer were relatively expensive.

Further, this commit improves the `Counter<A>` implementation
such that it now returns just a reference to the nonce
instead of allocating and copying a new one per fragment.