handle non-closed Enc/DecWriter in Drop impl.
This commit changes the behavior of the Drop implementation
for EncWriter and DecWriter such that the destructor panics
if the writer was not closed explicitly.

In general, a constructor should not panic. However, not doing
so when the writer hasn't been closed may cause silent data corruption.
In cryptographic code we cannot do the same optimistic approach
like e.g. std::io::BufWriter. Not calling `close()` is considered a
logic bug. Ideally, we would like to raise a compiler error when a
Enc/DecWriter is not closed but that implies that the compiler *proofs*
statically that the writer is *always* closed. Even though that may be
possible for the most cases (e.t. requiring complex analysis) it may
not be possible all the time... (Threads?!) - Anyway, if it turns out
that we can do this - we should!

However, for debugging purposes we may want to disable the panic to e.g.
debug a panic that occurs between creating an Enc/DecWriter and closing
it. (Dropping the writer causes another panic and we loose the
information from the first one...). Therefore, there is the feature
`must_close` which is enabled by default and disabling it only affects
debug builds.

Further, the close function is now annotated by the `must_use` attribute
since only after close you can tell whether i.e. decryption was
successful.