add benchmarks for EncWriter and DecWriter
This commit adds benchmarks for encryption
and decryption. With the `SIO_BUF_SIZE` env.
var. (in KiB) it is possible to customize use different
buffer sizes. If `SIO_BUF_SIZE` is not present the default
buffer size is used.

On important caveat is that decryption benchmarks only process
half the amount of data. The reason is that the data gets encrypted
first and then decrypted again. So, decryption benchmarks may not be
100% accurate.