check for panic during drop
This commit replaces the manual panic
tracking with `std::thread::panicking()`.

This eliminates the need for the `debug_panic`
feature which is also removed by this commit.

In addition this commit fixes the `write_buffer`
implementations to set the `errored` flag correctly.
In particular, the `errored` flag is now set when no
more nonces are available, en/decrypting fails or when
writing to the underlying writer fails.

A new set of unit tests ensures that `EncWriter` and `DecWriter`
behave as expected when not closed.